+++ tests/unit-tests.cpp
@@ -68,6 +68,7 @@ extern "C"
     int unit_test__storage__storage_property_string_check();
     int unit_test__storage__copy_string();
     int unit_test__storage_properties_set_chunking_props();
+    int unit_test__storage_properties_set_sharding_props();
     int unit_test__device_state_as_string__is_defined_for_all();
     int unit_test__device_kind_as_string__is_defined_for_all();
     int unit_test__sample_type_as_string__is_defined_for_all();
@@ -93,6 +94,7 @@ main()
         CASE(unit_test__storage__storage_property_string_check),
         CASE(unit_test__storage__copy_string),
         CASE(unit_test__storage_properties_set_chunking_props),
+        CASE(unit_test__storage_properties_set_sharding_props),
         CASE(unit_test__device_state_as_string__is_defined_for_all),
         CASE(unit_test__device_kind_as_string__is_defined_for_all),
         CASE(unit_test__sample_type_as_string__is_defined_for_all),